move theme provider to its own folder
